Show noobs how to pull request
