+++ src/app/views/login/page.jsx
@@ -2,8 +2,8 @@ import React from "react";
 
 export default function Login() {
   return (
-    <div className="flex justify-center items-center h-screen">
-      <form className="bg-white shadow-md rounded px-8 pt-6 pb-8 mb-4"></form>
+    <div className="flex justify-center items-center h-screen bg-black">
+      <form className="bg-blue-500 shadow-md rounded px-8 py-8"></form>
     </div>
   );
 }
